revision of data spooling code
Made Spooler more generic with data source choices set by an integer corresponding to class attributes that can be added to. Method data_source added to carry out the logic of deciding the source given a simple set of inputs. 'rfits' from ccd.py renamed to 'rhcam' in this file.