added routines to enforce particular data types
numpy promotes data types. e.g. if a float is added to an integer, the result is a float. Usually it's best to let it do this, but on output to disk one may want to reduce the number of bits used. This is supported with routines float32 and uint16 added to Windat, CCD and MCCD objects.