have added a trap for ctrl-C in grab so that temporary files can be deleted
grab has a mode set by temp in which it writes files to the local temp disc. These would get left if it was aborted using ctrl-C. Therefore I now trap these so that the files are deleted before leaving. This has a knock-on cosequence for 'makebias' that uses grab.