lots more additions to the hlogger code
Also, found some problems with the timing of runs on 2017-10-22 (and one on the previous night). Seems some extra bytes have crept in. The final 36 are timing bytes, but the origin of the extra bytes is a mystery. Example run is run0061 from 2017-10-22. It seems to have 196 "timing" bytes, so 160 bytes are unaccounted for.