significant change to the way apertures are moved
Now does not carry out a search on non-reference apertures, if there are some reference apertures. As a result the need for two search half widths has gone. Also changed genred to allow different values for the search half width and the profile fitting one as they are not closely related.