added a number of checks against bad reduce file input and fixed tkeep operation
It now checks for consistency between the reduce and aperture files after Matthew and Mark ran into a problem while observing. This should result in less opaque error messages. I also made the "tkeep" param behave in the same way as the max_xrange parameter in the ULTRACAM pipeline allowing you a shifting-window view of the last tkeep minutes of data