added a differential shift parameter.
If there is more than one reference aperture, this parameter is applied to test the consistency of the positions shifts from frame to frame. Since these are typically expected to be very similar, this is a sensitive way to check for problems in the reference positions. For instance, applied to run018 of an ULTRACAM run of 2017-05-04, the distribution of differential shifts in CCDs 1 and 2 was approx a 2D Maxwellian, with a mean of around 0.2 to 0.25 unbinned pixels. A differential shift limit of 2 then generated no false alarms in well over 2000 frames, but was easily sufficient to fix a problem caused by a cosmic ray near the third aperture on frame 17995.