removing fits.Header from ucam (first set of changes)
Thought I had done this, but this marks the start of removal of the astropy.io.fits.Header objects from ucam in the same way as accomplished for HiPERCAM. This was prompted by Paul Kerry's discovery of a bug with astropy v4.0 which in turn was related to the use of the astropy fits headers in ucam. Since we wanted to get rid of these anyway, I thought I might as well get on with it. Splitting off from the main development branch since it will take a while and some care needs to be taken.