small bug fix, changed to circular fit regions
I now implement a mask such that the fits are carried out over circular regions centred on the centre of the Window input. The regions are circular in terms of unbinned coordinates. I do this for symmetry rather than any speed saving (the calculations still proceed over the whole window).