nasty bug to do with parallel processing fixed
changes made to ccdaper inside moveApers which I had thought would propagate into rfile.aper were not at the end of parallel runs. The result was a big discontinuity in position at the end of each CCD group during parallel processing. This is a significant bug in the case of fit_alpha < 1 since the position is updated with some lag. psf_reduce will need checking for the same issue. The key change was to extract what had changed from the results and update the values outside the parallel loop ready for next time.

Signed-off-by: Tom Marsh <t.r.marsh@warwick.ac.uk>