Environment variable HIPERCAM_DEFAULT_SOURCE introduced
This is to allow the reset value of "source", the data source use by multiple routines to be set by the environment. The idea is that on the drpcs one really wants this to point to the servers, not to a local file. i.e.on the ULTRACAM and ULTRASPEC drpcs, set HIPERCAM_DEFAULT_SOURCE to ot "us", but on the HiPERCAM drpc, set it to "hs". Other users need not set this usually.